--- src/main/java/techcourse/myblog/UserInfo.java
@@ -1,4 +1,0 @@
-package techcourse.myblog;
-
-public interface UserInfo {
-}
