--- src/main/java/techcourse/myblog/domain/AuthenticationFailedException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.domain;
-
-public class AuthenticationFailedException extends RuntimeException {
-    public AuthenticationFailedException(String message) {
-        super(message);
-    }
-}
