+++ src/main/java/techcourse/myblog/service/dto/ArticleDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.User;
