+++ src/main/java/techcourse/myblog/support/argument/LoginUserHandlerMethodArgumentResolver.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web.support;
+package techcourse.myblog.support.argument;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -6,7 +6,6 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.web.LoginUser;
 
 import javax.servlet.http.HttpServletRequest;
