+++ src/main/java/techcourse/myblog/support/validation/UserInfo.java
@@ -1,0 +1,4 @@
+package techcourse.myblog.support.validation;
+
+public interface UserInfo {
+}
