+++ src/main/java/techcourse/myblog/web/controller/UserController.java
@@ -7,12 +7,12 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
-import techcourse.myblog.UserInfo;
+import techcourse.myblog.support.validation.UserInfo;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.dto.UserDto;
+import techcourse.myblog.service.dto.UserDto;
 import techcourse.myblog.service.UserReadService;
 import techcourse.myblog.service.UserWriteService;
-import techcourse.myblog.web.LoginUser;
+import techcourse.myblog.support.argument.LoginUser;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.groups.Default;
