+++ src/test/java/techcourse/myblog/service/CommentServiceTests.java
@@ -8,7 +8,7 @@ import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.Comment;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.domain.repository.CommentRepository;
-import techcourse.myblog.dto.CommentDto;
+import techcourse.myblog.service.dto.CommentDto;
 
 import java.util.Optional;
